Try and implement the par idea, need a simple button in the deal stage, have it appear when the cards overlap, pause the overlap animation, done that So now actually implement the par method First i need to find the total value of the deal Need to move the deal server side, all players are just dealing their own cards there is no persistency Only player 1 should deal, other players should receive the deal data At the minute the server has a player value, which is the current player but also a friendly players hashmap which is all players, it has a variable called id which will be the current player’s id, the player class has an index, which will be 0 1 or 2 based on if they are player 1 2 or 3
So to find out if the current player is player 1 just do MyServer.player.index
Player 1 will be 0, so only deal if we are player 0, otherwise receive data about that deal
The player index was not set correctly,
When we have all players emit the player ids and index to the server, the server can broadcast it to the other players,
This seems to work

Ok so the gotAllPlayers
Method will set off several methods which result in the players getting a unique index, it seems to display in the log that they have unique indexes, but when i go to the deal stage and call
MyServer.player.index == 0
It seems more than one can have index == 0

Only one ‘receives the deal’ and 2 ‘send deal’

Need to look into this next time

+++ core/src/com/gmail/andrewahughes/CardButton.java
@@ -72,13 +72,13 @@ public class CardButton  extends Actor {
     static byte dealAnimationSuit =-1;
     static byte dealAnimationPreviousSuit =0;
 
-    byte value;
+    int value;
 
     /*when the cards are dealt to the player, all that happens is this
     * variable is set to the player's index, which will be used in the
-    * touch logic, deafult is 0 so we know it doesn't have a player,
+    * touch logic, default is 0 so we know it doesn't have a player,
     * otherwise value will be 1, 2 or 3*/
-    byte playerIndex =0;
+    int playerIndex =0;
 
     /**constructor for triButton
      *
@@ -258,8 +258,7 @@ public class CardButton  extends Actor {
      * been set in setDealAnimationPosition method
      */
     public void moveToPositionAnimation(float percentage){
-        /*first we need to work out the correct position for this cardButton based
-        * on the playerIndex and the index - as well as the numberOfPlayers*/
+        /*first we need to work out the correct position for this cardButton */
 
         if(percentage<=0.5f){
             setX(getX()+((dealAnimationPositionX) - getX())*(percentage*2));
